Remove sensitive data from .gemrc
The idea is to have ~/.gemrc.local with credentials and ~/.gemrc for
all other configuration.

See:
https://www.soulcutter.com/articles/hiding-gemrc-credentials-in-dotfiles.html

Note: All sources need to be in ~/.gemrc.local, not just the ones
that require an API token. This is because settings are merged
all-or-nothing (https://stackoverflow.com/a/35051387/473467)

So ~/.gemrc.local needs to be:

    :sources:
    - https://rubygems.org/
    - https://TOKEN@rubygems.pkg.github.com/org/